feat: kakao.tsx 파일에 조건문 설정하여 무한 리디렉션 방지
